Unify the data table into one search-params hook
Collapses the server and client data tables into a single useDataTable that
keeps page, sort and filters in the URL (nuqs). Pass pageCount to run it
manual/server-side; omit it and the table sorts, filters and pages a local
array in memory from the same URL state — one hook, consumer's choice.

- Removes data-table-client. useDataTable now injects the per-variant
  in-memory filter functions and derives manual mode from pageCount.
- Encodes filter values by meta.variant (arrays for select / multiSelect /
  range / dateRange, scalars for text / number / date) so a multi-word
  text filter survives a URL round-trip.
- Simplifies dashboard-07 to a single in-memory table and drops the nested
  card (no card-in-card), with a cleaner header.
