+++ docs/conventions.md
@@ -105,9 +105,15 @@ heading rather than starting a new doc.
   table is one item per file, wired together with `registryDependencies`; it's
   showcased through the `dashboard-07` block rather than cramped component
   previews, since a table wants the full-width block surface.
-- **URL-state blocks wrap themselves in a nuqs adapter.** The `dashboard-07`
-  block's server table uses `useDataTable`, which syncs page/sort/filters to the
-  URL via nuqs, so the block wraps its content in `NuqsAdapter`
+- **One `useDataTable`, server or client.** The hook always keeps page, sort
+  and per-column filters in the URL (nuqs). Pass `pageCount` and it runs
+  manual (the server queries; `data` is the current page); omit it and the
+  table sorts, filters and pages the full `data` array in memory from the same
+  URL state. Filter values are encoded per `meta.variant` — arrays for
+  `select` / `multiSelect` / `range` / `dateRange`, scalars for `text` /
+  `number` / `date` — so a multi-word text filter survives a round-trip.
+- **URL-state blocks wrap themselves in a nuqs adapter.** `dashboard-07` uses
+  `useDataTable`, so it wraps its content in `NuqsAdapter`
   (`nuqs/adapters/next/app`). That's safe under `output: "export"` because
   block previews render client-side through `React.lazy` + `Suspense`
   (`registry-demos.tsx`), so `useSearchParams` is never called during the
