+++ registry.json
@@ -3441,7 +3441,7 @@
       "name": "dashboard-07",
       "type": "registry:block",
       "title": "Dashboard 7",
-      "description": "Accounts table section with the same data shown two ways: a client-side table that sorts, filters and pages in memory, and a server-driven table whose query lives in the URL. Faceted, text, range and date filters, sortable columns, row selection and pagination, framed in a card with tabs.",
+      "description": "Accounts table section: faceted, text, range and date filters, sortable columns, column visibility, row selection and pagination, with the whole view kept in the URL. Driven by the search-params useDataTable hook, here filtering and paging a local array in memory.",
       "categories": [
         "blocks",
         "dashboard"
@@ -3452,14 +3452,10 @@
       ],
       "registryDependencies": [
         "badge",
-        "card",
         "checkbox",
         "data-table",
-        "data-table-client",
         "data-table-column-header",
-        "data-table-parsers",
         "data-table-toolbar",
-        "tabs",
         "use-data-table"
       ],
       "files": [
@@ -3802,7 +3798,7 @@
       "name": "use-data-table",
       "type": "registry:ui",
       "title": "useDataTable",
-      "description": "Hook for the server-driven data table: TanStack Table in manual mode with pagination, sorting and filters synced to the URL via nuqs.",
+      "description": "Search-params-driven data table hook: keeps pagination, sorting and filters in the URL via nuqs. Pass pageCount to drive it server-side, or omit it to sort, filter and page a local array in memory — same hook, consumer's choice.",
       "categories": [
         "primitives"
       ],
@@ -3826,7 +3822,7 @@
       "name": "data-table",
       "type": "registry:ui",
       "title": "Data Table",
-      "description": "Renderer for the data table: takes a TanStack table instance and draws the header, rows, empty state and pagination. Compose it with useDataTable (server) or useDataTableClient, plus the toolbar and filters.",
+      "description": "Renderer for the data table: takes a TanStack table instance and draws the header, rows, empty state and pagination. Compose it with the useDataTable hook plus the toolbar and filters.",
       "categories": [
         "primitives"
       ],
@@ -3843,30 +3839,6 @@
           "path": "registry/hirael/ui/data-table.tsx",
           "type": "registry:ui",
           "target": "components/ui/data-table.tsx"
-        }
-      ]
-    },
-    {
-      "name": "data-table-client",
-      "type": "registry:ui",
-      "title": "useDataTableClient",
-      "description": "Hook for the client-side data table: TanStack Table that sorts, filters and pages a local array in memory, with a per-variant filter function wired from each column's meta. Ships a DataTableClient convenience component.",
-      "categories": [
-        "primitives"
-      ],
-      "dependencies": [
-        "@tanstack/react-table"
-      ],
-      "registryDependencies": [
-        "data-table",
-        "data-table-toolbar",
-        "data-table-utils"
-      ],
-      "files": [
-        {
-          "path": "registry/hirael/ui/data-table-client.tsx",
-          "type": "registry:ui",
-          "target": "components/ui/data-table-client.tsx"
         }
       ]
     }
