+++ registry/hirael/blocks/dashboard-07/dashboard-07.tsx
@@ -3,35 +3,14 @@
 
 import * as React from "react";
 import { NuqsAdapter } from "nuqs/adapters/next/app";
-import {
-  parseAsArrayOf,
-  parseAsInteger,
-  parseAsString,
-  useQueryState,
-  useQueryStates,
-} from "nuqs";
 import type { ColumnDef } from "@tanstack/react-table";
 
 import { Badge } from "@/registry/hirael/ui/badge";
-import {
-  Card,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "@/registry/hirael/ui/card";
 import { Checkbox } from "@/registry/hirael/ui/checkbox";
 import { DataTable } from "@/registry/hirael/ui/data-table";
 import { DataTableColumnHeader } from "@/registry/hirael/ui/data-table-column-header";
 import { DataTableToolbar } from "@/registry/hirael/ui/data-table-toolbar";
-import { getSortingStateParser } from "@/registry/hirael/ui/data-table-parsers";
-import { useDataTableClient } from "@/registry/hirael/ui/data-table-client";
 import { useDataTable } from "@/registry/hirael/ui/use-data-table";
-import {
-  Tabs,
-  TabsContent,
-  TabsList,
-  TabsTrigger,
-} from "@/registry/hirael/ui/tabs";
 
 type Account = {
   id: string;
@@ -106,10 +85,9 @@ const shortDate = new Intl.DateTimeFormat(undefined, {
   year: "numeric",
 });
 
-function useAccountColumns(filterable: Set<string>): ColumnDef<Account>[] {
-  return React.useMemo<ColumnDef<Account>[]>(() => {
-    const canFilter = (id: string) => filterable.has(id);
-    return [
+function useAccountColumns(): ColumnDef<Account>[] {
+  return React.useMemo<ColumnDef<Account>[]>(
+    () => [
       {
         id: "select",
         header: ({ table }) => (
@@ -148,7 +126,7 @@ function useAccountColumns(filterable: Set<string>): ColumnDef<Account>[] {
             </span>
           </div>
         ),
-        enableColumnFilter: canFilter("customer"),
+        enableColumnFilter: true,
         meta: {
           label: "Customer",
           placeholder: "Search customer…",
@@ -168,7 +146,7 @@ function useAccountColumns(filterable: Set<string>): ColumnDef<Account>[] {
                 : "secondary";
           return <Badge variant={tone}>{STATUS_LABELS[status]}</Badge>;
         },
-        enableColumnFilter: canFilter("status"),
+        enableColumnFilter: true,
         meta: {
           label: "Status",
           variant: "multiSelect",
@@ -181,10 +159,8 @@ function useAccountColumns(filterable: Set<string>): ColumnDef<Account>[] {
       {
         accessorKey: "plan",
         header: "Plan",
-        cell: ({ row }) => (
-          <span className="capitalize">{PLAN_LABELS[row.original.plan]}</span>
-        ),
-        enableColumnFilter: canFilter("plan"),
+        cell: ({ row }) => PLAN_LABELS[row.original.plan],
+        enableColumnFilter: true,
         meta: {
           label: "Plan",
           variant: "select",
@@ -197,7 +173,7 @@ function useAccountColumns(filterable: Set<string>): ColumnDef<Account>[] {
           <DataTableColumnHeader column={column} label="MRR" />
         ),
         cell: ({ row }) => currency.format(row.original.mrr),
-        enableColumnFilter: canFilter("mrr"),
+        enableColumnFilter: true,
         meta: { label: "MRR", variant: "range", range: [0, 1500], unit: "$" },
       },
       {
@@ -213,103 +189,23 @@ function useAccountColumns(filterable: Set<string>): ColumnDef<Account>[] {
         accessorKey: "signedUp",
         header: "Signed up",
         cell: ({ row }) => shortDate.format(row.original.signedUp),
-        enableColumnFilter: canFilter("signedUp"),
+        enableColumnFilter: true,
         meta: { label: "Signed up", variant: "date" },
       },
-    ];
-  }, [filterable]);
+    ],
+    [],
+  );
 }
 
-const CLIENT_FILTERS = new Set([
-  "customer",
-  "status",
-  "plan",
-  "mrr",
-  "signedUp",
-]);
-const SERVER_FILTERS = new Set(["status", "plan"]);
-const SERVER_COLUMN_IDS = [
-  "customer",
-  "status",
-  "plan",
-  "mrr",
-  "seats",
-  "signedUp",
-];
-
-function ClientAccountsTable() {
-  const columns = useAccountColumns(CLIENT_FILTERS);
-  const { table } = useDataTableClient({
+function AccountsTable() {
+  const columns = useAccountColumns();
+  const { table } = useDataTable({
     data: DATA,
     columns,
     getRowId: (row) => row.id,
-    initialState: { pagination: { pageIndex: 0, pageSize: 8 } },
-  });
-
-  return (
-    <DataTable table={table}>
-      <DataTableToolbar table={table} />
-    </DataTable>
-  );
-}
-
-function queryAccounts(
-  all: Account[],
-  q: {
-    page: number;
-    perPage: number;
-    sort: { id: string; desc: boolean }[];
-    status: string[] | null;
-    plan: string[] | null;
-  },
-) {
-  let rows = all;
-  if (q.status?.length) rows = rows.filter((r) => q.status?.includes(r.status));
-  if (q.plan?.length) rows = rows.filter((r) => q.plan?.includes(r.plan));
-
-  const sort = q.sort[0];
-  if (sort) {
-    rows = [...rows].sort((a, b) => {
-      const av = a[sort.id as keyof Account];
-      const bv = b[sort.id as keyof Account];
-      const cmp =
-        typeof av === "number" && typeof bv === "number"
-          ? av - bv
-          : String(av).localeCompare(String(bv));
-      return sort.desc ? -cmp : cmp;
-    });
-  }
-
-  const pageCount = Math.max(1, Math.ceil(rows.length / q.perPage));
-  const start = (q.page - 1) * q.perPage;
-  return { rows: rows.slice(start, start + q.perPage), pageCount };
-}
-
-function ServerAccountsTable() {
-  const columns = useAccountColumns(SERVER_FILTERS);
-
-  const [{ page, perPage, status, plan }] = useQueryStates({
-    page: parseAsInteger.withDefault(1),
-    perPage: parseAsInteger.withDefault(8),
-    status: parseAsArrayOf(parseAsString),
-    plan: parseAsArrayOf(parseAsString),
-  });
-  const [sort] = useQueryState(
-    "sort",
-    getSortingStateParser<Account>(new Set(SERVER_COLUMN_IDS)).withDefault([]),
-  );
-
-  const { rows, pageCount } = React.useMemo(
-    () => queryAccounts(DATA, { page, perPage, sort, status, plan }),
-    [page, perPage, sort, status, plan],
-  );
-
-  const { table } = useDataTable({
-    data: rows,
-    columns,
-    pageCount,
-    getRowId: (row) => row.id,
-    initialState: { pagination: { pageIndex: 0, pageSize: 8 } },
+    initialState: {
+      pagination: { pageIndex: 0, pageSize: 8 },
+    },
   });
 
   return (
@@ -326,29 +222,20 @@ export default function DashboardSeven() {
         data-slot="data-table-block"
         className="mx-auto w-full max-w-6xl p-4 sm:p-6"
       >
-        <Card className="gap-0 overflow-hidden p-0">
-          <CardHeader className="border-b p-6">
-            <CardTitle className="text-lg">Accounts</CardTitle>
-            <CardDescription>
-              {DATA.length} accounts. Filter, sort and page through the data —
-              client-side in memory, or server-driven with the query in the URL.
-            </CardDescription>
-          </CardHeader>
-          <Tabs defaultValue="client" className="gap-0">
-            <div className="border-b px-6 pt-4">
-              <TabsList>
-                <TabsTrigger value="client">Client-side</TabsTrigger>
-                <TabsTrigger value="server">Server-driven</TabsTrigger>
-              </TabsList>
-            </div>
-            <TabsContent value="client" className="p-6">
-              <ClientAccountsTable />
-            </TabsContent>
-            <TabsContent value="server" className="p-6">
-              <ServerAccountsTable />
-            </TabsContent>
-          </Tabs>
-        </Card>
+        <div className="mb-5 flex flex-wrap items-end justify-between gap-3">
+          <div className="space-y-1">
+            <h2 className="font-semibold text-xl tracking-tight">Accounts</h2>
+            <p className="text-muted-foreground text-sm">
+              Filter, sort and page through {DATA.length} accounts. The view
+              lives in the URL, so any filtered, sorted page is a shareable
+              link.
+            </p>
+          </div>
+          <Badge variant="secondary" className="font-normal">
+            {DATA.length} total
+          </Badge>
+        </div>
+        <AccountsTable />
       </section>
     </NuqsAdapter>
   );
